Refactor auth related methods from UserService to AuthService
