+++ src/services/public/UserService.ts
@@ -1,12 +1,7 @@
-import axios from 'axios';
-
-import { Cookie } from 'cookiejar';
-
 import { Extractors } from '../../collections/Extractors';
 import { RawAnalyticsGranularity, RawAnalyticsMetric } from '../../enums/raw/Analytics';
 import { ResourceType } from '../../enums/Resource';
 import { ProfileUpdateOptions } from '../../models/args/ProfileArgs';
-import { AuthCredential } from '../../models/auth/AuthCredential';
 import { Analytics } from '../../models/data/Analytics';
 import { BookmarkFolder } from '../../models/data/BookmarkFolder';
 import { CursoredData } from '../../models/data/CursoredData';
@@ -63,6 +58,13 @@ export class UserService extends FetcherService {
 		super(config);
 	}
 
+	/**
+	 * Gets the size in bytes of a base64 string.
+	 *
+	 * @param base64Data - The base64 data show size is required.
+	 *
+	 * @returns The size in bytes of the data.
+	 */
 	private _base64ByteSize(base64Data: string): number {
 		const paddingMatch = base64Data.match(/=+$/);
 		const paddingLength = paddingMatch ? paddingMatch[0].length : 0;
@@ -70,6 +72,13 @@ export class UserService extends FetcherService {
 		return (base64Data.length * 3) / 4 - paddingLength;
 	}
 
+	/**
+	 * Normalizes base64 data into just the raw base64 string.
+	 *
+	 * @param payload - The data to normalize.
+	 *
+	 * @returns The raw base64 part of the data.
+	 */
 	private _normalizeBase64(payload: string): string {
 		const trimmedPayload = payload.trim();
 		const lowerCasePayload = trimmedPayload.toLowerCase();
@@ -83,113 +92,6 @@ export class UserService extends FetcherService {
 		}
 
 		return trimmedPayload;
-	}
-
-	private _refreshApiKeyFromResponseCookies(setCookieHeader: string | string[] | undefined): void {
-		if (!this.config.apiKey || !setCookieHeader) {
-			return;
-		}
-
-		const requiredCookieNames = new Set(['auth_token', 'ct0', 'kdt', 'twid']);
-		const currentCookieString = AuthService.decodeCookie(this.config.apiKey);
-		const cookiePairs = this._splitSetCookieHeader(setCookieHeader);
-		const cookiesMap = new Map<string, string>();
-
-		for (const cookieEntry of currentCookieString.split(';')) {
-			const trimmedEntry = cookieEntry.trim();
-			const separatorIndex = trimmedEntry.indexOf('=');
-
-			if (!trimmedEntry || separatorIndex < 1) {
-				continue;
-			}
-
-			const key = trimmedEntry.slice(0, separatorIndex).trim();
-			const value = trimmedEntry.slice(separatorIndex + 1).trim();
-			if (!key || !value || !requiredCookieNames.has(key)) {
-				continue;
-			}
-
-			cookiesMap.set(key, value);
-		}
-
-		let hasUpdate = false;
-		for (const cookiePair of cookiePairs) {
-			const cookieValuePair = cookiePair.split(';', 1)[0]?.trim();
-			const separatorIndex = cookieValuePair?.indexOf('=') ?? -1;
-
-			if (!cookieValuePair || separatorIndex < 1) {
-				continue;
-			}
-
-			const key = cookieValuePair.slice(0, separatorIndex).trim();
-			const value = cookieValuePair.slice(separatorIndex + 1).trim();
-			if (!key || !value || !requiredCookieNames.has(key)) {
-				continue;
-			}
-
-			cookiesMap.set(key, value);
-			hasUpdate = true;
-		}
-
-		if (!hasUpdate || !cookiesMap.has('twid')) {
-			return;
-		}
-
-		let mergedCookieString = '';
-		for (const [key, value] of cookiesMap.entries()) {
-			mergedCookieString += `${key}=${value};`;
-		}
-
-		if (!mergedCookieString) {
-			return;
-		}
-
-		try {
-			this.config.apiKey = AuthService.encodeCookie(mergedCookieString);
-		} catch {
-			// Ignore cookie rotation errors and leave existing apiKey unchanged.
-		}
-	}
-
-	/**
-	 * Fetches a fresh ct0 (CSRF token) from Twitter by making a lightweight
-	 * authenticated request, then rotates the apiKey with the updated cookie.
-	 */
-	private async _refreshCsrfToken(): Promise<void> {
-		if (!this.config.apiKey) {
-			return;
-		}
-
-		try {
-			const cred = new AuthCredential(
-				AuthService.decodeCookie(this.config.apiKey)
-					.split(';')
-					.map((item) => new Cookie(item)),
-			);
-
-			const refreshResponse = await axios.get('https://x.com/i/api/1.1/account/verify_credentials.json', {
-				headers: {
-					...cred.toHeader(),
-					authorization:
-						'Bearer AAAAAAAAAAAAAAAAAAAAANRILgAAAAAAnNwIzUejRCOuH5E6I8xnZz4puTs%3D1Zv7ttfk8LF81IUq16cHjhLTvJu4FA33AGWWjCpTnA',
-				},
-				httpAgent: this.config.httpsAgent,
-				httpsAgent: this.config.httpsAgent,
-				validateStatus: () => true,
-			});
-
-			this._refreshApiKeyFromResponseCookies(refreshResponse.headers['set-cookie']);
-		} catch {
-			// Best-effort: if ct0 refresh fails, leave apiKey as-is
-		}
-	}
-
-	private _splitSetCookieHeader(setCookieHeader: string | string[]): string[] {
-		if (Array.isArray(setCookieHeader)) {
-			return setCookieHeader;
-		}
-
-		return setCookieHeader.split(/,(?=\s*[^;,]+=)/g);
 	}
 
 	private _validateBase64Payload(payload: string, fieldName: string): string {
@@ -487,14 +389,26 @@ export class UserService extends FetcherService {
 	public async changePassword(currentPassword: string, newPassword: string): Promise<boolean> {
 		const resource = ResourceType.USER_PASSWORD_CHANGE;
 
+		// Changing the password
 		const response = await this.requestWithResponse<IUserChangePasswordResponse>(resource, {
 			changePassword: { currentPassword, newPassword },
 		});
 
+		// Getting if password change was successful or not
 		const data = Extractors[resource](response.data) ?? false;
-		if (data) {
-			this._refreshApiKeyFromResponseCookies(response.headers['set-cookie']);
-			await this._refreshCsrfToken();
+
+		// If password change was successful
+		if (data === true) {
+			// Getting the new API key
+			const newApiKey = AuthService.getApiKeyFromReponse(response);
+
+			// If new API key is generated, update current API key
+			if (newApiKey !== undefined) {
+				this.config.apiKey = newApiKey;
+			}
+
+			// Getting the new CSRF token and updating current API key
+			await AuthService.refreshCsrfToken(this.config);
 		}
 
 		return data;
@@ -523,10 +437,12 @@ export class UserService extends FetcherService {
 			throw new Error('Username can only contain letters, numbers, and underscores');
 		}
 
+		// Changing the username
 		const response = await this.request<IUserSettingsResponse>(resource, {
 			username,
 		});
 
+		// Getting the updated username
 		const updatedUsername = Extractors[resource](response);
 
 		return updatedUsername?.toLowerCase() === username.toLowerCase();
